Give the todo window its own file, since the card outgrew the limit (#625)
The workbench arrived from main at 986 lines and the active-marker change took
it to 1014, over the 1000-line hard limit. The todo window is the part this
issue touches and it is self-contained: resolving field state, windowing a long
list, and deciding what each line says. It moves whole into a child module, the
way app_init and mouse_routing already split, so nothing about the public
surface changes.

The module doc says which of its two jobs is a truth claim and which is not.
The task state decides the checkbox and the active marker and comes from the
producer; the window start only decides which slice is on screen and asserts
nothing about the items in it.

The four new tests move to the sibling test file for the same reason: keeping
them in place would have pushed it past the size recommendation it exists to
respect.
