+++ project-plans/issue625-plan.md
@@ -109,6 +109,7 @@ ingress instead of rendering a confidently wrong checklist.
 |---|---|
 | New `MAX_TODO_STATE_BYTES` bound and specification §13 row | Every wire string in JSP/1 is bounded; an unbounded open-ended field would be the only hole |
 | Compliance fixtures/scenarios/traces edited wholesale | The retired field appears in each of them; a half-converted corpus would fail its own profiles |
+| `src/workbench_todo_window.rs` extracted from `src/workbench_view.rs` | Forced by the hard source-size gate, not chosen. The workbench arrived from main at 986 lines and the active-marker change pushed it to 1014, over the 1000 limit. The todo window is the cohesive piece this issue touches, so it moves whole into a `#[path]` child module — the pattern `app_init.rs` and `mouse_routing.rs` already use — with no public surface change. The two workbench test files were rebalanced for the same reason |
 | The status workbench's active marker (`src/workbench_view.rs` and its tests) | #640 landed the workbench on main mid-flight, carrying the exact derived marker the issue names: `▸` on the first item that merely was not finished. It is step 4 of the issue's own direction, it cannot compile against the amended domain type anyway, and leaving a guess beside an authoritative field would be a half-finished cutover |
 | ~~`src/app_input/prs_lifecycle_key_tests.rs` restored to a compiling state~~ | Withdrawn. Approved as an unrelated exception while `origin/main` at `6b6d9289` could not build its binary test target, then superseded: #645 fixed the same defect on main with `NavState::rooted`, which states the session's screen exactly instead of pushing it onto the dashboard. Integrating main resolved the conflict in main's favour, so the branch no longer changes this file at all |
